% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\name{add_sff_input}
\alias{add_sff_input}
\title{Add input for 'FormulaSFTerm'}
\usage{
add_sff_input(df, model, debug = FALSE)
}
\arguments{
\item{df}{the data frame}

\item{model}{an object of class \code{\link{TSModel}}}

\item{debug}{print debugging messages?}
}
\description{
Duplicates original data columns, adding new columns with the \code{_kg} and
\code{_ks} suffixes if these are missing.
}
