% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\name{create_jm_grid}
\alias{create_jm_grid}
\title{Create a dense grid until event time for each subject (for 'JointModel')}
\usage{
create_jm_grid(
  df_tte,
  df_lon,
  num_points = 30,
  id_var = "id",
  time_var = "t",
  even_spacing = FALSE
)
}
\arguments{
\item{df_tte}{time-to-event data frame}

\item{df_lon}{the longitudinal data frame}

\item{num_points}{number of grid points}

\item{id_var}{id variable}

\item{time_var}{time variable}

\item{even_spacing}{space grid points evenly?}
}
\description{
Create a dense grid until event time for each subject (for 'JointModel')
}
