% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terms-TermList.R
\name{create_termlist}
\alias{create_termlist}
\title{Create additive terms}
\usage{
create_termlist(formula, priors)
}
\arguments{
\item{formula}{the model formula}

\item{priors}{Prior configurations.}
}
\value{
An object of \code{R6} class \code{TermList}.
}
\description{
Create additive terms
}
