% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-FunctionDraws.R
\name{-.FunctionDraws}
\alias{-.FunctionDraws}
\title{Subtract}
\usage{
\method{-}{FunctionDraws}(f1, f2)
}
\arguments{
\item{f1}{component 1}

\item{f2}{component 2}
}
\value{
the difference \code{f1 - f2}
}
\description{
Subtract
}
