% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{example}
\alias{example}
\title{Run an example}
\usage{
example(num_bf = 32, scale_bf = 1.5, formula = "y ~ gp(x)", ...)
}
\arguments{
\item{num_bf}{Number of basis functions.}

\item{scale_bf}{Basis function domain scale.}

\item{formula}{The model formula.}

\item{...}{Other arguments to the \code{$fit()} method of
\code{\link{TSModel}}.}
}
\value{
An \code{\link{TSModelFit}} object.
}
\description{
Fits a model to simple simulated data.
}
