% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{example2}
\alias{example2}
\title{Run another example}
\usage{
example2(formula = "y ~ gp(time) + gp(time, arm)", ...)
}
\arguments{
\item{formula}{The model formula.}

\item{...}{Other arguments to the \code{$fit()} method of
\code{\link{LonModel}}.}
}
\value{
An \code{\link{LonModelFit}} object.
}
\description{
Fits a model to simple simulated data.
}
