% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\name{extend_df}
\alias{extend_df}
\title{Extend data frame}
\usage{
extend_df(df, t, time_var)
}
\arguments{
\item{df}{original data frame}

\item{t}{vector of new time values}

\item{time_var}{name of time variable}
}
\value{
new data frame
}
\description{
Extend data frame
}
