% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\name{extend_df2}
\alias{extend_df2}
\title{Extend data frame and add a constant continuous variable}
\usage{
extend_df2(df, t, time_var, x, x_var, id_var)
}
\arguments{
\item{df}{original data frame}

\item{t}{vector of new time values}

\item{time_var}{name of time variable}

\item{x}{value of other continuous variable (will be constant)}

\item{x_var}{name of other continuous variable}

\item{id_var}{name of id variable}
}
\value{
new data frame
}
\description{
Extend data frame and add a constant continuous variable
}
