% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-FunctionDraws.R
\name{FunctionDraws}
\alias{FunctionDraws}
\title{The 'FunctionDraws' class}
\description{
The 'FunctionDraws' class

The 'FunctionDraws' class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FunctionDraws-new}{\code{FunctionDraws$new()}}
\item \href{#method-FunctionDraws-num_draws}{\code{FunctionDraws$num_draws()}}
\item \href{#method-FunctionDraws-num_points}{\code{FunctionDraws$num_points()}}
\item \href{#method-FunctionDraws-get_name}{\code{FunctionDraws$get_name()}}
\item \href{#method-FunctionDraws-rename}{\code{FunctionDraws$rename()}}
\item \href{#method-FunctionDraws-get_input}{\code{FunctionDraws$get_input()}}
\item \href{#method-FunctionDraws-get_output}{\code{FunctionDraws$get_output()}}
\item \href{#method-FunctionDraws-as_data_frame}{\code{FunctionDraws$as_data_frame()}}
\item \href{#method-FunctionDraws-as_data_frame_long}{\code{FunctionDraws$as_data_frame_long()}}
\item \href{#method-FunctionDraws-quantiles_df}{\code{FunctionDraws$quantiles_df()}}
\item \href{#method-FunctionDraws-print}{\code{FunctionDraws$print()}}
\item \href{#method-FunctionDraws-plot}{\code{FunctionDraws$plot()}}
\item \href{#method-FunctionDraws-scale}{\code{FunctionDraws$scale()}}
\item \href{#method-FunctionDraws-add}{\code{FunctionDraws$add()}}
\item \href{#method-FunctionDraws-subtract}{\code{FunctionDraws$subtract()}}
\item \href{#method-FunctionDraws-log}{\code{FunctionDraws$log()}}
\item \href{#method-FunctionDraws-exp}{\code{FunctionDraws$exp()}}
\item \href{#method-FunctionDraws-zero_offset}{\code{FunctionDraws$zero_offset()}}
\item \href{#method-FunctionDraws-split_by_id}{\code{FunctionDraws$split_by_id()}}
\item \href{#method-FunctionDraws-zero_offset_by_id}{\code{FunctionDraws$zero_offset_by_id()}}
\item \href{#method-FunctionDraws-clone}{\code{FunctionDraws$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FunctionDraws-new"></a>}}
\if{latex}{\out{\hypertarget{method-FunctionDraws-new}{}}}
\subsection{Method \code{new()}}{
Create the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FunctionDraws$new(input, output, name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The data frame used as function input.}

\item{\code{output}}{An \code{rvar} containing draws of the function values
at \code{input} points.}

\item{\code{name}}{Function name.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FunctionDraws-num_draws"></a>}}
\if{latex}{\out{\hypertarget{method-FunctionDraws-num_draws}{}}}
\subsection{Method \code{num_draws()}}{
Get number of function draws
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FunctionDraws$num_draws()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FunctionDraws-num_points"></a>}}
\if{latex}{\out{\hypertarget{method-FunctionDraws-num_points}{}}}
\subsection{Method \code{num_points()}}{
Get number of input points.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FunctionDraws$num_points()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FunctionDraws-get_name"></a>}}
\if{latex}{\out{\hypertarget{method-FunctionDraws-get_name}{}}}
\subsection{Method \code{get_name()}}{
Get the function name.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FunctionDraws$get_name()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FunctionDraws-rename"></a>}}
\if{latex}{\out{\hypertarget{method-FunctionDraws-rename}{}}}
\subsection{Method \code{rename()}}{
Rename the function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FunctionDraws$rename(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{new name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FunctionDraws-get_input"></a>}}
\if{latex}{\out{\hypertarget{method-FunctionDraws-get_input}{}}}
\subsection{Method \code{get_input()}}{
Get the input data frame.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FunctionDraws$get_input()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FunctionDraws-get_output"></a>}}
\if{latex}{\out{\hypertarget{method-FunctionDraws-get_output}{}}}
\subsection{Method \code{get_output()}}{
Get the output \code{rvar}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FunctionDraws$get_output(as_matrix = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{as_matrix}}{Transform the \code{rvar} to a draws matrix
(see the \code{posterior} package).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FunctionDraws-as_data_frame"></a>}}
\if{latex}{\out{\hypertarget{method-FunctionDraws-as_data_frame}{}}}
\subsection{Method \code{as_data_frame()}}{
Get the input and output as \code{data.frame} with an \code{rvar}
column.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FunctionDraws$as_data_frame()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FunctionDraws-as_data_frame_long"></a>}}
\if{latex}{\out{\hypertarget{method-FunctionDraws-as_data_frame_long}{}}}
\subsection{Method \code{as_data_frame_long()}}{
Get the input and output as a long \code{data.frame} with each draw
stacked on top of each other.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FunctionDraws$as_data_frame_long()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FunctionDraws-quantiles_df"></a>}}
\if{latex}{\out{\hypertarget{method-FunctionDraws-quantiles_df}{}}}
\subsection{Method \code{quantiles_df()}}{
Create (a possibly filtered) data frame for ggplot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FunctionDraws$quantiles_df(
  ci_inner = 0.5,
  ci_outer = 0.8,
  filter_by = NULL,
  kept_vals = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ci_inner}}{inner credible interval}

\item{\code{ci_outer}}{outer credible interval}

\item{\code{filter_by}}{filtering variable}

\item{\code{kept_vals}}{value of the filter variable that are kept}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FunctionDraws-print"></a>}}
\if{latex}{\out{\hypertarget{method-FunctionDraws-print}{}}}
\subsection{Method \code{print()}}{
Print object description.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FunctionDraws$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FunctionDraws-plot"></a>}}
\if{latex}{\out{\hypertarget{method-FunctionDraws-plot}{}}}
\subsection{Method \code{plot()}}{
Plot using automatically selected aesthetics by default.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FunctionDraws$plot(
  x_var = "auto",
  group_by = "auto",
  color_by = "auto",
  facet_by = "auto",
  ribbon_alpha = 0.3,
  ci_inner = 0.5,
  ci_outer = 0.8,
  max_levels = 40,
  filter_by = NULL,
  kept_vals = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x_var}}{Name of the x-axis variable.}

\item{\code{group_by}}{Name of the factor used as group aesthetic.}

\item{\code{color_by}}{Name of the factor used as color and fill aesthetics.}

\item{\code{facet_by}}{Faceting factor.}

\item{\code{ribbon_alpha}}{Opacity of ribbon.}

\item{\code{ci_inner}}{Inner credible interval (ribbon).}

\item{\code{ci_outer}}{Outer credible interval (ribbon).}

\item{\code{max_levels}}{Maximum number of levels to facet by.}

\item{\code{filter_by}}{Factor to filter the rows by.}

\item{\code{kept_vals}}{Values of the factor that are not filtered out.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FunctionDraws-scale"></a>}}
\if{latex}{\out{\hypertarget{method-FunctionDraws-scale}{}}}
\subsection{Method \code{scale()}}{
Scale the draws as \code{f_new = f*a + b}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FunctionDraws$scale(a = 1, b = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{a}}{multiplier}

\item{\code{b}}{added constant}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FunctionDraws-add"></a>}}
\if{latex}{\out{\hypertarget{method-FunctionDraws-add}{}}}
\subsection{Method \code{add()}}{
Add to another \code{FunctionDraws}. You can also use \code{f1 + f2}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FunctionDraws$add(f)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{f}}{A \code{FunctionDraws} object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FunctionDraws-subtract"></a>}}
\if{latex}{\out{\hypertarget{method-FunctionDraws-subtract}{}}}
\subsection{Method \code{subtract()}}{
Subtract another \code{FunctionDraws}. You can also use \code{f1 - f2}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FunctionDraws$subtract(f)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{f}}{A \code{FunctionDraws} object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FunctionDraws-log"></a>}}
\if{latex}{\out{\hypertarget{method-FunctionDraws-log}{}}}
\subsection{Method \code{log()}}{
Take log of function draws.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FunctionDraws$log()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FunctionDraws-exp"></a>}}
\if{latex}{\out{\hypertarget{method-FunctionDraws-exp}{}}}
\subsection{Method \code{exp()}}{
Exponentiate function draws.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FunctionDraws$exp()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FunctionDraws-zero_offset"></a>}}
\if{latex}{\out{\hypertarget{method-FunctionDraws-zero_offset}{}}}
\subsection{Method \code{zero_offset()}}{
Offset every function draw so that they start from zero
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FunctionDraws$zero_offset()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a new object of class \code{\link{FunctionDraws}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FunctionDraws-split_by_id"></a>}}
\if{latex}{\out{\hypertarget{method-FunctionDraws-split_by_id}{}}}
\subsection{Method \code{split_by_id()}}{
Split to groups
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FunctionDraws$split_by_id(id_var)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id_var}}{Name of the subject identifier factor.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a list of objects of class \code{\link{FunctionDraws}},
one for each group/id
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FunctionDraws-zero_offset_by_id"></a>}}
\if{latex}{\out{\hypertarget{method-FunctionDraws-zero_offset_by_id}{}}}
\subsection{Method \code{zero_offset_by_id()}}{
Offset every function draw for each subject so that they start from zero
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FunctionDraws$zero_offset_by_id(id_var)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id_var}}{Name of the subject identifier factor.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a new object of class \code{\link{FunctionDraws}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FunctionDraws-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FunctionDraws-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FunctionDraws$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
