% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models-Likelihood.R
\name{Likelihood}
\alias{Likelihood}
\title{Observation model class (R6 class)}
\description{
Observation model class (R6 class)

Observation model class (R6 class)
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Likelihood-clone}{\code{Likelihood$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Likelihood-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Likelihood-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Likelihood$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
