% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-rvar.R
\name{new_draws}
\alias{new_draws}
\title{Reset values of given parameters in 'rvars' list to a constant value}
\usage{
new_draws(rvars, names, value = 0)
}
\arguments{
\item{rvars}{An object of class \code{draws_rvars}.}

\item{names}{A character vector of parameter names.}

\item{value}{Value for the parameters.}
}
\value{
An updated object of class \code{draws_rvars}.
}
\description{
Reset values of given parameters in 'rvars' list to a constant value
}
