% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction-search.R
\name{plot_pp_elpd}
\alias{plot_pp_elpd}
\title{Plot result of projection predictive forward search (ELPD)}
\usage{
plot_pp_elpd(res, elpd_ref)
}
\arguments{
\item{res}{The list returned by \code{\link{pp_forward_search}}}

\item{elpd_ref}{ELPD of reference model (estimate and SE)}
}
\description{
Plot result of projection predictive forward search (ELPD)
}
