% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction-search.R
\name{plot_pp_elpd_diff}
\alias{plot_pp_elpd_diff}
\title{Plot result of projection predictive forward search (ELPD rel diff)}
\usage{
plot_pp_elpd_diff(res)
}
\arguments{
\item{res}{The list returned by \code{\link{pp_forward_search}}}
}
\description{
Plot result of projection predictive forward search (ELPD rel diff)
}
