% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction-hs_smooth.R
\name{Predict.matrix.hs.smooth}
\alias{Predict.matrix.hs.smooth}
\title{prediction method function for the `hs' smooth class}
\usage{
\method{Predict.matrix}{hs.smooth}(object, data)
}
\description{
prediction method function for the `hs' smooth class
}
