% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction-hs_smooth.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{smooth.construct}
\alias{Predict.matrix}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{mgcv}{\code{\link[mgcv]{Predict.matrix}}, \code{\link[mgcv]{smooth.construct}}}
}}

