% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-simulate.R
\name{simulate}
\alias{simulate}
\title{Data simulation for model reduction experiments}
\usage{
simulate(n, rho, rel_true, sigma)
}
\arguments{
\item{n}{number of data points}

\item{rho}{correlation}

\item{rel_true}{true relevances}

\item{sigma}{noise magnitude}
}
\description{
Data simulation for model reduction experiments
}
