% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction-hs_smooth.R
\name{smooth.construct.hs.smooth.spec}
\alias{smooth.construct.hs.smooth.spec}
\title{hilbert space GP}
\usage{
\method{smooth.construct}{hs.smooth.spec}(object, data, knots)
}
\description{
hilbert space GP
}
