% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models-TSModel.R
\name{stancode_ts}
\alias{stancode_ts}
\title{Can be used to just generate Stan code for a TS model}
\usage{
stancode_ts(formula, print = TRUE, ...)
}
\arguments{
\item{formula}{model formula}

\item{print}{Should the code be printed?}

\item{...}{Arguments passed to the \code{$create_stancode()} method
of \code{\link{TSModel}}.}
}
\value{
Stan code as character string (invisibly)
}
\description{
Can be used to just generate Stan code for a TS model
}
