% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models-StanModel.R
\name{StanCodeCreator}
\alias{StanCodeCreator}
\title{Class for objects that can create parts of 'Stan' code}
\description{
Class for objects that can create parts of 'Stan' code

Class for objects that can create parts of 'Stan' code
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StanCodeCreator-string}{\code{StanCodeCreator$string()}}
\item \href{#method-StanCodeCreator-print}{\code{StanCodeCreator$print()}}
\item \href{#method-StanCodeCreator-stanfiles_functions}{\code{StanCodeCreator$stanfiles_functions()}}
\item \href{#method-StanCodeCreator-stancode_functions}{\code{StanCodeCreator$stancode_functions()}}
\item \href{#method-StanCodeCreator-stancode_data}{\code{StanCodeCreator$stancode_data()}}
\item \href{#method-StanCodeCreator-stancode_tdata}{\code{StanCodeCreator$stancode_tdata()}}
\item \href{#method-StanCodeCreator-stancode_pars}{\code{StanCodeCreator$stancode_pars()}}
\item \href{#method-StanCodeCreator-stancode_tpars}{\code{StanCodeCreator$stancode_tpars()}}
\item \href{#method-StanCodeCreator-stancode_model}{\code{StanCodeCreator$stancode_model()}}
\item \href{#method-StanCodeCreator-stancode_gq}{\code{StanCodeCreator$stancode_gq()}}
\item \href{#method-StanCodeCreator-clone}{\code{StanCodeCreator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StanCodeCreator-string"></a>}}
\if{latex}{\out{\hypertarget{method-StanCodeCreator-string}{}}}
\subsection{Method \code{string()}}{
Description as a string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanCodeCreator$string()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StanCodeCreator-print"></a>}}
\if{latex}{\out{\hypertarget{method-StanCodeCreator-print}{}}}
\subsection{Method \code{print()}}{
Print info.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanCodeCreator$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Nothing.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StanCodeCreator-stanfiles_functions"></a>}}
\if{latex}{\out{\hypertarget{method-StanCodeCreator-stanfiles_functions}{}}}
\subsection{Method \code{stanfiles_functions()}}{
Stan files where to get the functions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanCodeCreator$stanfiles_functions()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StanCodeCreator-stancode_functions"></a>}}
\if{latex}{\out{\hypertarget{method-StanCodeCreator-stancode_functions}{}}}
\subsection{Method \code{stancode_functions()}}{
Generate 'Stan' code for the functions block.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanCodeCreator$stancode_functions()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StanCodeCreator-stancode_data"></a>}}
\if{latex}{\out{\hypertarget{method-StanCodeCreator-stancode_data}{}}}
\subsection{Method \code{stancode_data()}}{
Generate 'Stan' code for the data block.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanCodeCreator$stancode_data(datanames = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{datanames}}{Names of input data sets.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StanCodeCreator-stancode_tdata"></a>}}
\if{latex}{\out{\hypertarget{method-StanCodeCreator-stancode_tdata}{}}}
\subsection{Method \code{stancode_tdata()}}{
Generate 'Stan' code for the transformed data block.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanCodeCreator$stancode_tdata(datanames = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{datanames}}{Names of input data sets.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StanCodeCreator-stancode_pars"></a>}}
\if{latex}{\out{\hypertarget{method-StanCodeCreator-stancode_pars}{}}}
\subsection{Method \code{stancode_pars()}}{
Generate 'Stan' code for the parameters block.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanCodeCreator$stancode_pars()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StanCodeCreator-stancode_tpars"></a>}}
\if{latex}{\out{\hypertarget{method-StanCodeCreator-stancode_tpars}{}}}
\subsection{Method \code{stancode_tpars()}}{
Generate 'Stan' code for the transformed parameters block.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanCodeCreator$stancode_tpars(datanames = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{datanames}}{Names of input data sets.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StanCodeCreator-stancode_model"></a>}}
\if{latex}{\out{\hypertarget{method-StanCodeCreator-stancode_model}{}}}
\subsection{Method \code{stancode_model()}}{
Generate 'Stan' code for the model block.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanCodeCreator$stancode_model()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StanCodeCreator-stancode_gq"></a>}}
\if{latex}{\out{\hypertarget{method-StanCodeCreator-stancode_gq}{}}}
\subsection{Method \code{stancode_gq()}}{
Generate 'Stan' code for the generated quantities block.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanCodeCreator$stancode_gq()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StanCodeCreator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StanCodeCreator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanCodeCreator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
