% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models-StanModel.R
\name{StanModel}
\alias{StanModel}
\title{Abstract for objects that can generate entire 'Stan' models}
\description{
Abstract for objects that can generate entire 'Stan' models

Abstract for objects that can generate entire 'Stan' models
}
\section{Super class}{
\code{\link[lgpr2:StanCodeCreator]{lgpr2::StanCodeCreator}} -> \code{StanModel}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StanModel-new}{\code{StanModel$new()}}
\item \href{#method-StanModel-get_stanmodel}{\code{StanModel$get_stanmodel()}}
\item \href{#method-StanModel-create_stancode}{\code{StanModel$create_stancode()}}
\item \href{#method-StanModel-create_stanmodel}{\code{StanModel$create_stanmodel()}}
\item \href{#method-StanModel-compile}{\code{StanModel$compile()}}
\item \href{#method-StanModel-clone}{\code{StanModel$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="lgpr2" data-topic="StanCodeCreator" data-id="print"><a href='../../lgpr2/html/StanCodeCreator.html#method-StanCodeCreator-print'><code>lgpr2::StanCodeCreator$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lgpr2" data-topic="StanCodeCreator" data-id="stancode_data"><a href='../../lgpr2/html/StanCodeCreator.html#method-StanCodeCreator-stancode_data'><code>lgpr2::StanCodeCreator$stancode_data()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lgpr2" data-topic="StanCodeCreator" data-id="stancode_functions"><a href='../../lgpr2/html/StanCodeCreator.html#method-StanCodeCreator-stancode_functions'><code>lgpr2::StanCodeCreator$stancode_functions()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lgpr2" data-topic="StanCodeCreator" data-id="stancode_gq"><a href='../../lgpr2/html/StanCodeCreator.html#method-StanCodeCreator-stancode_gq'><code>lgpr2::StanCodeCreator$stancode_gq()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lgpr2" data-topic="StanCodeCreator" data-id="stancode_model"><a href='../../lgpr2/html/StanCodeCreator.html#method-StanCodeCreator-stancode_model'><code>lgpr2::StanCodeCreator$stancode_model()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lgpr2" data-topic="StanCodeCreator" data-id="stancode_pars"><a href='../../lgpr2/html/StanCodeCreator.html#method-StanCodeCreator-stancode_pars'><code>lgpr2::StanCodeCreator$stancode_pars()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lgpr2" data-topic="StanCodeCreator" data-id="stancode_tdata"><a href='../../lgpr2/html/StanCodeCreator.html#method-StanCodeCreator-stancode_tdata'><code>lgpr2::StanCodeCreator$stancode_tdata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lgpr2" data-topic="StanCodeCreator" data-id="stancode_tpars"><a href='../../lgpr2/html/StanCodeCreator.html#method-StanCodeCreator-stancode_tpars'><code>lgpr2::StanCodeCreator$stancode_tpars()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lgpr2" data-topic="StanCodeCreator" data-id="stanfiles_functions"><a href='../../lgpr2/html/StanCodeCreator.html#method-StanCodeCreator-stanfiles_functions'><code>lgpr2::StanCodeCreator$stanfiles_functions()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lgpr2" data-topic="StanCodeCreator" data-id="string"><a href='../../lgpr2/html/StanCodeCreator.html#method-StanCodeCreator-string'><code>lgpr2::StanCodeCreator$string()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StanModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-StanModel-new}{}}}
\subsection{Method \code{new()}}{
Create model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanModel$new(compile = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{compile}}{Should the 'Stan' model code be created and compiled.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StanModel-get_stanmodel"></a>}}
\if{latex}{\out{\hypertarget{method-StanModel-get_stanmodel}{}}}
\subsection{Method \code{get_stanmodel()}}{
Get the underlying 'Stan' model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanModel$get_stanmodel()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StanModel-create_stancode"></a>}}
\if{latex}{\out{\hypertarget{method-StanModel-create_stancode}{}}}
\subsection{Method \code{create_stancode()}}{
Create the 'Stan' model code for the model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanModel$create_stancode(autoformat = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{autoformat}}{Should automatic formatting be attempted?}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A string.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StanModel-create_stanmodel"></a>}}
\if{latex}{\out{\hypertarget{method-StanModel-create_stanmodel}{}}}
\subsection{Method \code{create_stanmodel()}}{
Create and compile the 'Stan' model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanModel$create_stanmodel(dir = tempdir())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dir}}{Path to directory where to store the \code{.stan} file.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A 'CmdStanR' model.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StanModel-compile"></a>}}
\if{latex}{\out{\hypertarget{method-StanModel-compile}{}}}
\subsection{Method \code{compile()}}{
Create and compile the 'Stan' model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanModel$compile(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed to \code{create_stanmodel()}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The updated model object (invisibly).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StanModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StanModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
