% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{stanmodel_functions}
\alias{stanmodel_functions}
\title{Create a 'Stan' model with only given functions in it}
\usage{
stanmodel_functions(stan_file, ...)
}
\arguments{
\item{stan_file}{Path to the Stan file relative to the directory
\code{inst/functions}, without the \code{.stan} suffix.}

\item{...}{Optional arguments to \code{cmdstanr::cmdstan_model()}.}
}
\description{
Create a 'Stan' model with only given functions in it
}
