% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fits-StanModelFit.R
\name{StanModelFit}
\alias{StanModelFit}
\title{The Fit class}
\description{
The Fit class

The Fit class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StanModelFit-get_data}{\code{StanModelFit$get_data()}}
\item \href{#method-StanModelFit-datanames}{\code{StanModelFit$datanames()}}
\item \href{#method-StanModelFit-get_model}{\code{StanModelFit$get_model()}}
\item \href{#method-StanModelFit-new}{\code{StanModelFit$new()}}
\item \href{#method-StanModelFit-get_stan_fit}{\code{StanModelFit$get_stan_fit()}}
\item \href{#method-StanModelFit-get_stan_data}{\code{StanModelFit$get_stan_data()}}
\item \href{#method-StanModelFit-draws}{\code{StanModelFit$draws()}}
\item \href{#method-StanModelFit-log_z_pars}{\code{StanModelFit$log_z_pars()}}
\item \href{#method-StanModelFit-loglik}{\code{StanModelFit$loglik()}}
\item \href{#method-StanModelFit-gq}{\code{StanModelFit$gq()}}
\item \href{#method-StanModelFit-clone}{\code{StanModelFit$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StanModelFit-get_data"></a>}}
\if{latex}{\out{\hypertarget{method-StanModelFit-get_data}{}}}
\subsection{Method \code{get_data()}}{
Get original data used to fit the model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanModelFit$get_data(dataname = "LON")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataname}}{Name of dataset.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StanModelFit-datanames"></a>}}
\if{latex}{\out{\hypertarget{method-StanModelFit-datanames}{}}}
\subsection{Method \code{datanames()}}{
Get names of stored data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanModelFit$datanames()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StanModelFit-get_model"></a>}}
\if{latex}{\out{\hypertarget{method-StanModelFit-get_model}{}}}
\subsection{Method \code{get_model()}}{
Get model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanModelFit$get_model(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of model. Only has effect for
\code{\link{JointModelFit}} objects.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StanModelFit-new"></a>}}
\if{latex}{\out{\hypertarget{method-StanModelFit-new}{}}}
\subsection{Method \code{new()}}{
Create model fit object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanModelFit$new(model, stan_fit, datasets, stan_data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{A \code{\link{StanModel}} object.}

\item{\code{stan_fit}}{The created 'Stan' fit.}

\item{\code{datasets}}{Original data sets.}

\item{\code{stan_data}}{The created 'Stan' data. Stored mainly for easier
debugging.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StanModelFit-get_stan_fit"></a>}}
\if{latex}{\out{\hypertarget{method-StanModelFit-get_stan_fit}{}}}
\subsection{Method \code{get_stan_fit()}}{
Get the underlying 'Stan' fit object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanModelFit$get_stan_fit()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StanModelFit-get_stan_data"></a>}}
\if{latex}{\out{\hypertarget{method-StanModelFit-get_stan_data}{}}}
\subsection{Method \code{get_stan_data()}}{
Get the underlying 'Stan' data object (for debugging).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanModelFit$get_stan_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StanModelFit-draws"></a>}}
\if{latex}{\out{\hypertarget{method-StanModelFit-draws}{}}}
\subsection{Method \code{draws()}}{
Extract draws as \code{rvar}s
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanModelFit$draws(name = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Param/quantity name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StanModelFit-log_z_pars"></a>}}
\if{latex}{\out{\hypertarget{method-StanModelFit-log_z_pars}{}}}
\subsection{Method \code{log_z_pars()}}{
Full names of parameters that start with 'log_z_'.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanModelFit$log_z_pars()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StanModelFit-loglik"></a>}}
\if{latex}{\out{\hypertarget{method-StanModelFit-loglik}{}}}
\subsection{Method \code{loglik()}}{
Extract log likelihood as 'rvars'.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanModelFit$loglik()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StanModelFit-gq"></a>}}
\if{latex}{\out{\hypertarget{method-StanModelFit-gq}{}}}
\subsection{Method \code{gq()}}{
Generate quantities using the fit.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanModelFit$gq(stan_data = NULL, fitted_params = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{stan_data}}{Full 'Stan' input list.}

\item{\code{fitted_params}}{Argument to \code{generate_quantities()}.}

\item{\code{...}}{Other arguments to \code{generate_quantities()}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StanModelFit-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StanModelFit-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StanModelFit$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
