% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{term_to_code}
\alias{term_to_code}
\title{Name of a term in the formula syntax to its Stan code name}
\usage{
term_to_code(term)
}
\arguments{
\item{term}{A string.}
}
\description{
Name of a term in the formula syntax to its Stan code name
}
