% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testdata}
\alias{testdata}
\title{Simulated longitudinal test data}
\format{
A data frame with 176 rows and 7 columns:
\describe{
  \item{id}{Subject id, factor}
  \item{arm}{Trial arm, factor}
  \item{sex}{Sex, factor}
  \item{time}{Measurement time, numeric}
  \item{weight}{Weight, numeric}
  \item{y}{Response variable, numeric}
  \item{f_true}{True signal}
}
}
\usage{
testdata
}
\description{
Simulated longitudinal test data
}
\concept{built-in datasets}
\keyword{datasets}
