% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\name{transform_df}
\alias{transform_df}
\title{Transform a data frame using a LonModel}
\usage{
transform_df(model, df)
}
\arguments{
\item{model}{the model}

\item{df}{original data frame}
}
\value{
new data frame
}
\description{
Transform a data frame using a LonModel
}
