% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{example2}
\alias{example2}
\title{Run an example}
\usage{
example2(
  num_bf = 24,
  scale_bf = 1.5,
  formula = "y ~ gp(time) + gp(time,arm) + gp(time,id)",
  ...
)
}
\arguments{
\item{num_bf}{Number of basis functions.}

\item{scale_bf}{Basis function domain scale.}

\item{formula}{The model formula.}

\item{...}{Other arguments to the \code{$fit()} method of
\code{\link{TSModel}}.}
}
\value{
An \code{\link{TSModelFit}} object.
}
\description{
Fits a model to \code{testdata}.
}
