% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models-Likelihood.R
\name{GaussianLikelihood}
\alias{GaussianLikelihood}
\title{Observation model class (R6 class)}
\description{
Observation model class (R6 class)

Observation model class (R6 class)
}
\section{Super class}{
\code{\link[lgpr2:Likelihood]{lgpr2::Likelihood}} -> \code{GaussianLikelihood}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{prior_sigma}}{Prior for noise magnitude parameter.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GaussianLikelihood-new}{\code{GaussianLikelihood$new()}}
\item \href{#method-GaussianLikelihood-stancode_data}{\code{GaussianLikelihood$stancode_data()}}
\item \href{#method-GaussianLikelihood-stancode_pars}{\code{GaussianLikelihood$stancode_pars()}}
\item \href{#method-GaussianLikelihood-stancode_tpars}{\code{GaussianLikelihood$stancode_tpars()}}
\item \href{#method-GaussianLikelihood-stancode_model}{\code{GaussianLikelihood$stancode_model()}}
\item \href{#method-GaussianLikelihood-stancode_gq}{\code{GaussianLikelihood$stancode_gq()}}
\item \href{#method-GaussianLikelihood-clone}{\code{GaussianLikelihood$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GaussianLikelihood-new"></a>}}
\if{latex}{\out{\hypertarget{method-GaussianLikelihood-new}{}}}
\subsection{Method \code{new()}}{
Create model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GaussianLikelihood$new(prior_sigma = "lognormal(1, 1)")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{prior_sigma}}{Prior for noise magnitude parameter.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GaussianLikelihood-stancode_data"></a>}}
\if{latex}{\out{\hypertarget{method-GaussianLikelihood-stancode_data}{}}}
\subsection{Method \code{stancode_data()}}{
Stan code for data block
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GaussianLikelihood$stancode_data(stanname_y, dataname)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{stanname_y}}{Name of y variable in Stan code}

\item{\code{dataname}}{suffix}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GaussianLikelihood-stancode_pars"></a>}}
\if{latex}{\out{\hypertarget{method-GaussianLikelihood-stancode_pars}{}}}
\subsection{Method \code{stancode_pars()}}{
Stan code for parameters block
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GaussianLikelihood$stancode_pars()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GaussianLikelihood-stancode_tpars"></a>}}
\if{latex}{\out{\hypertarget{method-GaussianLikelihood-stancode_tpars}{}}}
\subsection{Method \code{stancode_tpars()}}{
Stan code for transformed parameters block
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GaussianLikelihood$stancode_tpars(stanname_y, dataname)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{stanname_y}}{Name of y variable in Stan code}

\item{\code{dataname}}{suffix}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GaussianLikelihood-stancode_model"></a>}}
\if{latex}{\out{\hypertarget{method-GaussianLikelihood-stancode_model}{}}}
\subsection{Method \code{stancode_model()}}{
Stan code for model block
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GaussianLikelihood$stancode_model()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{stanname_y}}{Name of y variable in Stan code}

\item{\code{dataname}}{suffix}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GaussianLikelihood-stancode_gq"></a>}}
\if{latex}{\out{\hypertarget{method-GaussianLikelihood-stancode_gq}{}}}
\subsection{Method \code{stancode_gq()}}{
Stan code for the generated quantities block
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GaussianLikelihood$stancode_gq(y_name, stanname_y)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y_name}}{name of y variable}

\item{\code{stanname_y}}{Name of y variable in Stan code}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GaussianLikelihood-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GaussianLikelihood-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GaussianLikelihood$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
