% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fits-LonModelFit.R
\name{LonModelFit}
\alias{LonModelFit}
\title{The 'LonModelFit' class}
\description{
The 'LonModelFit' class

The 'LonModelFit' class
}
\section{Super class}{
\code{\link[lgpr2:StanModelFit]{lgpr2::StanModelFit}} -> \code{LonModelFit}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{term_confs}}{The term configurations used.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LonModelFit-new}{\code{LonModelFit$new()}}
\item \href{#method-LonModelFit-print}{\code{LonModelFit$print()}}
\item \href{#method-LonModelFit-stan_variables}{\code{LonModelFit$stan_variables()}}
\item \href{#method-LonModelFit-diagnose}{\code{LonModelFit$diagnose()}}
\item \href{#method-LonModelFit-function_draws}{\code{LonModelFit$function_draws()}}
\item \href{#method-LonModelFit-plot}{\code{LonModelFit$plot()}}
\item \href{#method-LonModelFit-predict}{\code{LonModelFit$predict()}}
\item \href{#method-LonModelFit-predict_time}{\code{LonModelFit$predict_time()}}
\item \href{#method-LonModelFit-noise_amount}{\code{LonModelFit$noise_amount()}}
\item \href{#method-LonModelFit-comp_vars}{\code{LonModelFit$comp_vars()}}
\item \href{#method-LonModelFit-relevances}{\code{LonModelFit$relevances()}}
\item \href{#method-LonModelFit-rank_terms}{\code{LonModelFit$rank_terms()}}
\item \href{#method-LonModelFit-reduce}{\code{LonModelFit$reduce()}}
\item \href{#method-LonModelFit-project}{\code{LonModelFit$project()}}
\item \href{#method-LonModelFit-clone}{\code{LonModelFit$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="lgpr2" data-topic="StanModelFit" data-id="datanames"><a href='../../lgpr2/html/StanModelFit.html#method-StanModelFit-datanames'><code>lgpr2::StanModelFit$datanames()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lgpr2" data-topic="StanModelFit" data-id="draws"><a href='../../lgpr2/html/StanModelFit.html#method-StanModelFit-draws'><code>lgpr2::StanModelFit$draws()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lgpr2" data-topic="StanModelFit" data-id="get_data"><a href='../../lgpr2/html/StanModelFit.html#method-StanModelFit-get_data'><code>lgpr2::StanModelFit$get_data()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lgpr2" data-topic="StanModelFit" data-id="get_model"><a href='../../lgpr2/html/StanModelFit.html#method-StanModelFit-get_model'><code>lgpr2::StanModelFit$get_model()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lgpr2" data-topic="StanModelFit" data-id="get_stan_data"><a href='../../lgpr2/html/StanModelFit.html#method-StanModelFit-get_stan_data'><code>lgpr2::StanModelFit$get_stan_data()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lgpr2" data-topic="StanModelFit" data-id="get_stan_fit"><a href='../../lgpr2/html/StanModelFit.html#method-StanModelFit-get_stan_fit'><code>lgpr2::StanModelFit$get_stan_fit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lgpr2" data-topic="StanModelFit" data-id="gq"><a href='../../lgpr2/html/StanModelFit.html#method-StanModelFit-gq'><code>lgpr2::StanModelFit$gq()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lgpr2" data-topic="StanModelFit" data-id="loglik"><a href='../../lgpr2/html/StanModelFit.html#method-StanModelFit-loglik'><code>lgpr2::StanModelFit$loglik()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lgpr2" data-topic="StanModelFit" data-id="loo_estimate"><a href='../../lgpr2/html/StanModelFit.html#method-StanModelFit-loo_estimate'><code>lgpr2::StanModelFit$loo_estimate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="lgpr2" data-topic="StanModelFit" data-id="num_draws"><a href='../../lgpr2/html/StanModelFit.html#method-StanModelFit-num_draws'><code>lgpr2::StanModelFit$num_draws()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LonModelFit-new"></a>}}
\if{latex}{\out{\hypertarget{method-LonModelFit-new}{}}}
\subsection{Method \code{new()}}{
Create model fit object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LonModelFit$new(model, stan_fit, datasets, stan_data, term_confs)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{A \code{\link{StanModel}} object.}

\item{\code{stan_fit}}{The created 'Stan' fit.}

\item{\code{datasets}}{Original data (list of data frames).}

\item{\code{stan_data}}{The created 'Stan' data. Stored mainly for easier
debugging.}

\item{\code{term_confs}}{The term configurations used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LonModelFit-print"></a>}}
\if{latex}{\out{\hypertarget{method-LonModelFit-print}{}}}
\subsection{Method \code{print()}}{
Print object description.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LonModelFit$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LonModelFit-stan_variables"></a>}}
\if{latex}{\out{\hypertarget{method-LonModelFit-stan_variables}{}}}
\subsection{Method \code{stan_variables()}}{
Stan variable names.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LonModelFit$stan_variables(all = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{all}}{Return all Stan variable names. If \code{FALSE}, returns only
ones that are not large transformed parameters.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LonModelFit-diagnose"></a>}}
\if{latex}{\out{\hypertarget{method-LonModelFit-diagnose}{}}}
\subsection{Method \code{diagnose()}}{
A diagnostic summary.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LonModelFit$diagnose()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LonModelFit-function_draws"></a>}}
\if{latex}{\out{\hypertarget{method-LonModelFit-function_draws}{}}}
\subsection{Method \code{function_draws()}}{
Extract one component as a \code{FunctionDraws} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LonModelFit$function_draws(component = "h", input_vars = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{component}}{Name of the component (term). Valid names can
be checked using \code{$model$term_names()}. Alternatively, can be
an integer that is the component index.}

\item{\code{input_vars}}{Column names in the data. Only has effect if
getting a single component. If \code{NULL}, determined automatically
from the component.}

\item{\code{dataname}}{name of data set used to evaluate the function}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LonModelFit-plot"></a>}}
\if{latex}{\out{\hypertarget{method-LonModelFit-plot}{}}}
\subsection{Method \code{plot()}}{
Plot fit or predictive distribution
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LonModelFit$plot(
  f_reference = NULL,
  plot_y = TRUE,
  predictive = TRUE,
  f_ref_name = "the true signal",
  filter_by = NULL,
  kept_vals = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{f_reference}}{Reference function to plot against the fit.}

\item{\code{plot_y}}{Should y data be plotted?}

\item{\code{predictive}}{Should this plot the predictive distribution for
new observations? Otherwise the inferred signal is plotted.}

\item{\code{f_ref_name}}{Name of the reference function.}

\item{\code{filter_by}}{Factor to filter the rows by.}

\item{\code{kept_vals}}{Values of the factor that are not filtered out.}

\item{\code{...}}{Arguments passed to the plot method of
\code{\link{FunctionDraws}}.
Compute predictions at test points}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LonModelFit-predict"></a>}}
\if{latex}{\out{\hypertarget{method-LonModelFit-predict}{}}}
\subsection{Method \code{predict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LonModelFit$predict(data = NULL, fitted_params = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{The input data frame of test points.}

\item{\code{fitted_params}}{Parameters or model fit.}

\item{\code{...}}{Other arguments to \code{generate_quantities()}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new fit object.
Compute predictions at given time points
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LonModelFit-predict_time"></a>}}
\if{latex}{\out{\hypertarget{method-LonModelFit-predict_time}{}}}
\subsection{Method \code{predict_time()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LonModelFit$predict_time(
  t_test,
  t_var = "t",
  y_test = NULL,
  t_var_copy = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{t_test}}{The test time points}

\item{\code{t_var}}{Name of the time variable}

\item{\code{y_test}}{Test y}

\item{\code{t_var_copy}}{Names of other continuous covariates to which
\code{t_test} should be copied.}

\item{\code{...}}{Arguments passed to \code{$predict()}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new fit object
Estimate amount of noise
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LonModelFit-noise_amount"></a>}}
\if{latex}{\out{\hypertarget{method-LonModelFit-noise_amount}{}}}
\subsection{Method \code{noise_amount()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LonModelFit$noise_amount()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A tibble with num of rows equal to number of draws. Values
are between 0 (no noise) and 1 (data is only noise).
Estimate component variances for each draw
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LonModelFit-comp_vars"></a>}}
\if{latex}{\out{\hypertarget{method-LonModelFit-comp_vars}{}}}
\subsection{Method \code{comp_vars()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LonModelFit$comp_vars()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A matrix with n_draws rows and n_comps columns
Estimate component relevances
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LonModelFit-relevances"></a>}}
\if{latex}{\out{\hypertarget{method-LonModelFit-relevances}{}}}
\subsection{Method \code{relevances()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LonModelFit$relevances(rvar = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rvar}}{return an \code{rvar} vector?}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LonModelFit-rank_terms"></a>}}
\if{latex}{\out{\hypertarget{method-LonModelFit-rank_terms}{}}}
\subsection{Method \code{rank_terms()}}{
Rank model terms based on relevance

Reduce model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LonModelFit$rank_terms()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LonModelFit-reduce"></a>}}
\if{latex}{\out{\hypertarget{method-LonModelFit-reduce}{}}}
\subsection{Method \code{reduce()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LonModelFit$reduce(thresh = 0.95)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{thresh}}{Threshold for explained variance.
Project to submodel}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LonModelFit-project"></a>}}
\if{latex}{\out{\hypertarget{method-LonModelFit-project}{}}}
\subsection{Method \code{project()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LonModelFit$project(
  term_inds = NULL,
  draw_inds = NULL,
  eval_mode = TRUE,
  B = 30
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{term_inds}}{Which terms to include in submodel?}

\item{\code{draw_inds}}{Which posterior draws to use. If \code{NULL}, 30
draws are taken randomly.}

\item{\code{eval_mode}}{Model evaluation mode?}

\item{\code{B}}{number of basis functions}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LonModelFit-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LonModelFit-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LonModelFit$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
