% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction-search.R
\name{plot_pp_elpd}
\alias{plot_pp_elpd}
\title{Plot result of projection predictive forward search (ELPD)}
\usage{
plot_pp_elpd(res, fit, ...)
}
\arguments{
\item{res}{The list returned by \code{\link{pp_forward_search}}}

\item{fit}{Reference model fit object.}

\item{...}{Arguments passed to \code{fit$loo_estimate()}.}
}
\description{
Plot result of projection predictive forward search (ELPD)
}
