% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction-search.R
\name{plot_pp_pexp}
\alias{plot_pp_pexp}
\title{Plot result of projection predictive forward search (p_explained)}
\usage{
plot_pp_pexp(res, thresh = 0.95)
}
\arguments{
\item{res}{The list returned by \code{\link{pp_forward_search}}}

\item{thresh}{A horizontal line.}
}
\description{
Plot result of projection predictive forward search (p_explained)
}
