% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction-search.R
\name{pp_forward_search}
\alias{pp_forward_search}
\title{Run projection predictive forward search}
\usage{
pp_forward_search(fit_ref, path = NULL, num_steps = NULL)
}
\arguments{
\item{fit_ref}{The reference model fit.}

\item{path}{Pre-defined search path. If \code{NULL}, path is
not pre-defined.}

\item{num_steps}{Number of steps to run. If \code{NULL}, the search stops
when predictive performance of the submodel is close to that of reference
model.}
}
\description{
Run projection predictive forward search
}
