% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction-search.R
\name{pp_forward_search}
\alias{pp_forward_search}
\title{Run projection predictive forward search}
\usage{
pp_forward_search(
  fit_ref,
  path = NULL,
  num_steps = NULL,
  S = 30,
  S_eval = 100,
  B = 24,
  random_idx = NULL
)
}
\arguments{
\item{fit_ref}{The reference model fit.}

\item{path}{Pre-defined search path. If \code{NULL}, path is
not pre-defined. These indices are same as the index of the term in
\code{fit_ref$get_model()$term_names()}.}

\item{num_steps}{Number of steps to run. If \code{NULL}, the search stops
when predictive performance of the submodel is close to that of reference
model.}

\item{S}{Number of draws to project when selecting next step.}

\item{S_eval}{Number of draws to project when evaluating step.}

\item{B}{number of basis functions}

\item{random_idx}{index of the "random effect" term}
}
\description{
Run projection predictive forward search
}
